move base RODSConn class methods to RC_base
